backend/requirements.txt
aiosqlite==0.19.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.10
django-allauth==0.61.1
django-cors-headers==4.3.1
django-filter==24.1
django-qrcode==0.3
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pypng==0.20220715.0
python-barcode==0.15.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
qrcode==7.4.2
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1